+++ src/Services/Admin/Settings/SettingsSanitizationSchema.php
@@ -16,14 +16,14 @@ class SettingsSanitizationSchema
     public static function getRules(): array
     {
         return [
-            PluginSettings::SECTION_ALERTS           => [
+            PluginSettings::SECTION_ALERTS        => [
                 PluginSettings::ALERT_TRAFFIC_DROP  => 'bool',
                 PluginSettings::ALERT_TRAFFIC_SPIKE => 'bool',
                 PluginSettings::ALERT_404_SPIKE     => 'bool',
                 PluginSettings::ALERT_BOT_SPIKE     => 'bool',
                 PluginSettings::ALERT_BOT_DROP      => 'bool',
             ],
-            PluginSettings::SECTION_THRESHOLDS => [
+            PluginSettings::SECTION_THRESHOLDS    => [
                 PluginSettings::MIN_WEEKLY_AVG          => 'int',
                 PluginSettings::MIN_PAGEVIEWS_FOR_ALERT => 'int',
                 PluginSettings::TRAFFIC_SPIKE_PERCENT   => 'int',
@@ -31,6 +31,13 @@ class SettingsSanitizationSchema
                 PluginSettings::ERROR_404_SPIKE_PERCENT => 'int',
                 PluginSettings::BOT_SPIKE_PERCENT       => 'int',
                 PluginSettings::BOT_DROP_PERCENT        => 'int',
+            ],
+            PluginSettings::SECTION_NOTIFICATIONS => [
+                PluginSettings::ENABLE_DAILY_DIGEST    => 'bool',
+                PluginSettings::DIGEST_RECIPIENT_EMAIL => 'email',
+                PluginSettings::DIGEST_INCLUDE_TRAFFIC => 'bool',
+                PluginSettings::DIGEST_INCLUDE_404     => 'bool',
+                PluginSettings::DIGEST_INCLUDE_BOT     => 'bool',
             ],
         ];
     }